// src/models/Alumno.ts
export interface Alumno {
  id: number; // ID único del alumno
  nombre: string; // Nombre del alumno
  apellido: string; // Apellido del alumno
  edad: number; // Edad del alumno
}

/*En resumen, este es un objeto de interfaz de TypeScript que define las propiedades de un alumno, 
incluyendo su ID, nombre, apellido y edad.*/